Simplify the Power ZEROCHK evaluator
Previously, the ZEROCHK evaluator in the Power codegen had a lot of
special casing to try to perform comparisons directly into condition
registers. Since functionality has now been added to perform this
optimization in a more general manner, this code can now be replaced
with a simple call to evaluateToConditionRegister.

Signed-off-by: Ben Thomas <ben@benthomas.ca>